+++ packages/broker/src/broker.ts
@@ -21,10 +21,6 @@ export interface Broker {
     stop: () => Promise<unknown>
 }
 
-const getNameDescription = (name: string|undefined, id: string) => {
-    return (name !== undefined) ? `${name} (id=${id})` : id
-}
-
 export const createBroker = async (config: Config): Promise<Broker> => {
     validateConfig(config, BROKER_CONFIG_SCHEMA)
     validateClientConfig(config.client)
@@ -69,7 +65,7 @@ export const createBroker = async (config: Config): Promise<Broker> => {
 
             logger.info(`Welcome to the Streamr Network. Your node's generated name is ${Protocol.generateMnemonicFromAddress(brokerAddress)}.`)
             logger.info(`View your node in the Network Explorer: https://streamr.network/network-explorer/nodes/${encodeURIComponent(nodeId)}`)
-            logger.info(`Network node ${getNameDescription(config.client.network?.name, nodeId)} running`)
+            logger.info(`Network node ${nodeId} running`)
             logger.info(`Ethereum address ${brokerAddress}`)
             logger.info(`Tracker Configuration: ${
                 config.client.network?.trackers ? JSON.stringify(config.client.network?.trackers) : 'default'